Removes theme
The website was unpublished as I tried to set the repo to be private.
After that, the website remained offline even when I published it again.
A hack said to add a theme that will update the _config.yml [1]. It
worked and afterwards, this commit removes the theme and restores the
original _config.yml.

[1]: https://stackoverflow.com/a/48592535